refactor(api): split riskScores.js controller into per-responsibility modules (#599)
riskScores.js had grown to 610 lines (audit finding C1). Split it by
responsibility into focused routers under routes/riskScores/:

  list.js   — GET /api/risk-scores (summary) + the per-type paginated lists
  entity.js — GET /api/risk-scores/:type/:id + PUT/DELETE .../override
  shared.js — useSql / db + riskTableExists / parseJsonColumns + TEMPORAL_FILTER

riskScores.js is now a barrel that composes the two routers, so app.js and
every public path are unchanged. Pure code move.

Grandfathered cyclomatic entries relocated (no increase): [37,27] ->
list [37] + entity [27] (no cognitive units were over threshold).

No new tests needed — the three existing riskScores test files cover the
moved code at 86%+ diff coverage. Full API suite green (1758 tests);
coverage above every floor; complexity + jscpd + OpenAPI-drift gates pass.
