+++ .ci/complexity-baseline.json
@@ -68,9 +68,11 @@
       25,
       24
     ],
-    "app/api/src/routes/riskScores.js": [
-      37,
+    "app/api/src/routes/riskScores/entity.js": [
       27
+    ],
+    "app/api/src/routes/riskScores/list.js": [
+      37
     ],
     "app/api/src/routes/tags/crud.js": [
       23
