+++ changes/split-fat-route-controllers-599.md
@@ -5,3 +5,4 @@
 - Internal maintainability: split the large tags API controller into focused modules (tag CRUD + assignment, and the users/groups/entity-tags list endpoints) behind a thin barrel. No functional change — every tags/users/groups endpoint behaves exactly as before.
 - Internal maintainability: split the large recent-changes API controller into focused modules (the recent-changes panels and the timeline endpoints, sharing common label-resolution helpers) behind a thin barrel, and added tests covering the event-building paths. No functional change — every recent-changes and timeline endpoint behaves exactly as before.
 - Internal maintainability: split the large contexts API controller into focused modules (context reads, context create/update/sync/delete, and membership management) behind a thin barrel. No functional change — every contexts endpoint behaves exactly as before.
+- Internal maintainability: split the large risk-scores API controller into focused modules (the summary/list endpoints and the single-entity detail + analyst-override endpoints) behind a thin barrel. No functional change — every risk-scores endpoint behaves exactly as before.
